backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
h11==0.14.0
httptools==0.6.1
idna==3.10
motor==3.3.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1